moved abstractResources from core into features, removed some comments and unnecessary imports
