+++ smart_farm_frontend/src/features/abstractResource/AbstractResource.ts
