+++ smart_farm_frontend/src/features/abstractResource/EnergyResource.ts
