+++ smart_farm_frontend/src/features/abstractResource/abstractExport.ts
