+++ smart_farm_frontend/src/features/abstractResource/types.ts
